+++ fcomb-docker-distribution/src/main/scala/io/fcomb/docker/distribution/server/api/WebhooksHandler.scala
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2016 fcomb. <https://fcomb.io>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.fcomb.docker.distribution.server.api
+
+import akka.http.scaladsl.model.HttpRequest
+import akka.http.scaladsl.model.headers.Accept
+import akka.http.scaladsl.server.Directives._
+import cats.data.Validated
+import io.fcomb.models.docker.distribution.{ImageWebhooksPutRequest, ImageWebhooksResponse}
+import io.fcomb.persist.docker.distribution.ImageWebhooksRepo
+import io.fcomb.server.CirceSupport._
+import io.circe.generic.auto._
+import io.fcomb.server.AuthenticationDirectives._
+import io.fcomb.server.ImageDirectives._
+import io.fcomb.models.acl.Action
+
+object WebhooksHandler {
+  def getWebhooks(imageName: String) =
+    authenticateUserBasic { user =>
+      extractMaterializer { implicit mat =>
+        optionalHeaderValueByType[Accept]() { acceptOpt =>
+          imageByNameWithAcl(imageName, user.getId(), Action.Read) { image =>
+            onSuccess(ImageWebhooksRepo.findByImageId(image.getId())) {
+              case Some(webhook) => complete(ImageWebhooksResponse(imageName, Seq(webhook.url)))
+              case _             => complete(ImageWebhooksResponse(imageName, Seq()))
+            }
+          }
+        }
+      }
+    }
+
+  def putWebhooks(imageName: String)(implicit req: HttpRequest) =
+    authenticateUserBasic { user =>
+      extractMaterializer { implicit mat =>
+        import mat.executionContext
+        imageByNameWithAcl(imageName, user.getId(), Action.Write) { image =>
+          entity(as[ImageWebhooksPutRequest]) { putRequest =>
+            onSuccess(ImageWebhooksRepo.upsert(image.getId(), putRequest.url)) {
+              case Validated.Valid(upserted) =>
+                complete(ImageWebhooksResponse(imageName, Seq(upserted.url)))
+              case _ => complete(ImageWebhooksResponse(imageName, Seq()))
+            }
+          }
+        }
+      }
+    }
+
+}
